fix: redux was breaking build, clean up, add letter to redux
