+++ gatsby-browser.js
@@ -1,9 +1,2 @@
-/**
- * Implement Gatsby's Browser APIs in this file.
- *
- * See: https://www.gatsbyjs.org/docs/browser-apis/
- */
-
-// You can delete this file if you're not using it
-import reduxWrapper from './redux-wrapper';
+import reduxWrapper from './src/redux/wrap-with-provider';
 export const wrapRootElement = reduxWrapper;
