+++ package.json
@@ -47,8 +47,9 @@
     "react": "^16.8.4",
     "react-dom": "^16.8.4",
     "react-helmet": "^5.2.0",
+    "react-quill": "^1.3.3",
     "react-redux": "^6.0.1",
-    "react-quill": "^1.3.3"
+    "redux": "^4.0.1"
   },
   "devDependencies": {
     "enzyme": "^3.9.0",
