+++ src/pages/compose.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { Link } from 'gatsby';
+import { connect } from 'react-redux';
 
 //components
 import Layout from '../components/layout';
@@ -7,24 +8,18 @@ import Label from '../components/label';
 import Button from '../components/button';
 import TextEditor from '../components/text-editor';
 
-//redux actions
-import { onClick } from '../redux/actions/compose';
+// Redux
+import { bindActionCreators } from 'redux';
+import actions from '../redux/actions/compose';
+const mapStateToProps = ({ compose }) => ({ ...compose });
+const mapDispatchToProps = dispatch => bindActionCreators(actions, dispatch);
 
-//redux config
-import { connect } from 'react-redux';
-const mapStateToProps = state => ({ composeProps: state.composeReducer });
-const mapDispatchToProps = dispatch => ({
-  onClick: () => dispatch(onClick())
-});
-
-
-const ComposePage = () => (
+const ComposePage = ({ onLetterChange, letterHTML }) => (
   <Layout>
     <Label cursive className="text-muted">
       Your Letter
     </Label>
-    <TextEditor onChange={value => console.log(value)} />
-
+    <TextEditor onChange={onLetterChange} value={letterHTML} />
     <div className="pt-5 d-flex justify-content-end">
       <Link to="/preview">
         <Button disabled primary text="Preview" />
@@ -33,12 +28,7 @@ const ComposePage = () => (
   </Layout>
 );
 
-//connecting view -> this gets exported and redered as the component
 export default connect(
   mapStateToProps,
   mapDispatchToProps
 )(ComposePage);
-
-
-
-
