+++ src/redux/TYPES/compose.js
@@ -1,0 +1,3 @@
+export default {
+  EDIT_LETTER: 'EDIT_LETTER'
+};
