--- src/redux/action-types/compose/index.js
@@ -1,1 +1,0 @@
-export const UPDATE_CHECKBOX = 'UPDATE_CHECKBOX';
