+++ src/redux/actions/compose.js
@@ -1,0 +1,8 @@
+import TYPES from '../TYPES/compose';
+
+export default {
+  onLetterChange: value => ({
+    type: TYPES.EDIT_LETTER,
+    payload: value
+  })
+};
