--- src/redux/actions/compose/index.js
@@ -1,5 +1,0 @@
-import { UPDATE_CHECKBOX } from '../../action-types/compose';
-
-export const onClick = () => ({
-  type: UPDATE_CHECKBOX
-});
