+++ src/redux/reducers/index.js
@@ -1,0 +1,6 @@
+import { combineReducers } from 'redux';
+
+// Reducers
+import compose from './compose';
+
+export default combineReducers({ compose });
