--- src/redux/reducers/reducers-combined.js
@@ -1,4 +1,0 @@
-import { combineReducers } from 'redux';
-import composeReducer from './compose/index';
-
-export default combineReducers({ composeReducer });
