+++ src/redux/wrap-with-provider.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import { Provider } from 'react-redux';
+import createStore from './store';
+
+export default ({ element }) => {
+  const store = createStore();
+  return <Provider store={store}>{element}</Provider>;
+};
